tunes/tester.rb renamed and gutted to just be about sandbox testers
