+++ spaceship/lib/spaceship/tunes/tunes.rb
@@ -22,7 +22,7 @@ require 'spaceship/tunes/build_details'
 require 'spaceship/tunes/build_train'
 require 'spaceship/tunes/device_type'
 require 'spaceship/tunes/app_trailer'
-require 'spaceship/tunes/tester'
+require 'spaceship/tunes/sandbox_tester'
 require 'spaceship/tunes/app_details'
 require 'spaceship/tunes/pricing_tier'
 require 'spaceship/tunes/recovery_device'
